Forgot to add files to Makefile and functions to .h Also rearranged files and functions again.
Known errors:
When executing a command that should have some kind of input (wc, sort) it goes infinite
 (as is supposed to). The problem is when CTRL-Cing, it creates new line but it
 repeats the prompt:

	minishell$>sort
	^C
	minishell$>minishell$>
